+++ ft_strnstr.c
@@ -14,8 +14,8 @@
 
 char	*ft_strnstr(const char *haystack, const char *needle, size_t len)
 {
-	int		hi;
-	int		ni;
+	size_t	hi;
+	size_t	ni;
 	char	*haystack_tmp;
 
 	hi = 0;
@@ -24,17 +24,17 @@ char	*ft_strnstr(const char *haystack, const char *needle, size_t len)
 	haystack_tmp = (char *)haystack;
 	if (ft_strlen(needle) == 0)
 		return ((char *)haystack);
-	while (hi < len && hi < ft_strlen(haystack))
+	while (hi < len && hi < (size_t)ft_strlen(haystack))
 	{
 		ni = 0;
 		while (needle[ni] != '\0' && hi + ni < len
-			&& hi + ni <= ft_strlen(haystack))
+			&& hi + ni <= (size_t)ft_strlen(haystack))
 		{
 			if (haystack[hi + ni] != needle[ni])
 				break ;
 			ni++;
 		}
-		if (ni == ft_strlen(needle))
+		if (ni == (size_t)ft_strlen(needle))
 			return (&haystack_tmp[hi]);
 		hi++;
 	}
